+++ CHANGELOG.md
@@ -1,23 +1,47 @@
+# 3.3.0
+## Added
+### Basic Nodes
+- Added a Load Image Node
+### Logic Nodes
+- Added a Logic Compare
+## Changed
+### Basic Nodes
+- Slightly refactored the Save Image node
+### Prompt Styler Node
+- Slightly refactored the Prompt Styler node
+### Math Nodes
+- Added the Constant as an output
+- Slightly refactored the Math node
+***
+
 # 3.2.0
 ## Added
 ### Combination Nodes
 - Added a Combine String Node
+***
+
 # 3.1.2
-## Changes
+## Changed
 ### README
 - Just some cleanup
+***
+
 # 3.1.1
-## Changes
+## Changed
 ### Basic Nodes
 - Fixed the order of the Prompt combinder's combining
 ### Math Nodes
 - Capitalized "Round"
+***
+
 # 3.1.0
-## Changes
+## Changed
 ### Math Node
 - Added "round" as an operation to the Math Operation Node
 - Changed the order of the Math node inputs to make more sense in my head
 - Changed "Number" to "Constant" in Math Node
+***
+
 # 3.0.0
 ## Added
 ### Conversion Nodes
@@ -25,44 +49,51 @@
 - String to Integer
 ### Math Nodes
 Added Math Nodes
-## Changes
+## Changed
 ### Multiple Nodes
 - Changed "int" to "integer" throughout the code just because I'm insane
 - Some more code cleanup as I'm learning new things
----
+***
+
 # 2.2.0
-## Changes
+## Changed
 ### Basic Nodes
 - Added a Face section to the Prompt combiner
 - Added outputs for each section of the Prompt Combiner
----
+***
+
 # 2.1.0
-## Changes
+## Changed
 ### Basic Nodes
 - Added a few blank lines to improve readability when logging prompts from prompt_styler
----
+***
+
 # 2.0.0
 ## Added
 ### Basic Nodes
 - Save Image Node
-## Changes
+## Changed
 - Some Cleanup and commenting for clarity
 - Super important correction on the publish.yml, definitely not hiding my misspelling of "correction"
----
+***
+
 # 1.1.1
-## Changes
+## Changed
 - Just changed the version ID in order to make the Comfy Registry update
----
+***
+
 # 1.1.0
 ## Added
 ### Basic Nodes
 - Remade the prompt styler node from scratch
----
+***
+
 # 1.0.1
 ## Removed
 ### Prompt Stylers
 - This was copied this code from https://github.com/TripleHeadedMonkey/ComfyUI_MileHighStyler and https://github.com/twri/sdxl_prompt_styler, thank you very much to them for the inspiration.
----
+***
+
 # 1.0.0
 ## Added
 ### Basic Nodes